# src/main/java/cn/edu/xidian/tafei_mall/controller/CartController.java
package cn.edu.xidian.tafei_mall.controller;

import cn.edu.xidian.tafei_mall.model.vo.*;
import org.springframework.web.bind.annotation.*;

import java.util.Map;

@RestController
@RequestMapping("/api/cart")
public class CartController {

    @GetMapping
    public ResponseEntity<?> getCartContents() {
        // Implement get cart contents logic
        return ResponseEntity.ok("获取成功");
    }

    @PostMapping("/items")
    public ResponseEntity<?> addItemToCart(@RequestBody CartItemAdd cartItemAdd) {
        // Implement add item to cart logic
        return ResponseEntity.status(HttpStatus.CREATED).body("添加成功");
    }

    @PutMapping("/items/{itemId}")
    public ResponseEntity<?> updateCartItem(@PathVariable String itemId, @RequestBody Map<String, Integer> update) {
        // Implement update cart item logic
        return ResponseEntity.ok("数量更新成功");
    }

    @DeleteMapping("/items/{itemId}")
    public ResponseEntity<?> deleteCartItem(@PathVariable String itemId) {
        // Implement delete cart item logic
        return ResponseEntity.noContent().build();
    }
}