# src/main/java/cn/edu/xidian/tafei_mall/controller/ProductController.java
package cn.edu.xidian.tafei_mall.controller;

import cn.edu.xidian.tafei_mall.model.vo.*;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/products")
public class ProductController {

    @GetMapping("/search")
    public ResponseEntity<?> searchProducts(@RequestParam String keyword, @RequestParam(defaultValue = "1") int page, @RequestParam(defaultValue = "10") int limit) {
        // Implement product search logic
        return ResponseEntity.ok("搜索成功");
    }

    @GetMapping("/{productId}")
    public ResponseEntity<?> getProductDetails(@PathVariable String productId) {
        // Implement get product details logic
        return ResponseEntity.ok("获取成功");
    }
}