# src/main/java/cn/edu/xidian/tafei_mall/model/vo/AddressUpdate.java
package cn.edu.xidian.tafei_mall.model.vo;


import lombok.Data;

@Data
public class AddressUpdate {
    private String address;
    private String city;
    private String postalCode;

    // Getters and Setters
}