# src/main/java/cn/edu/xidian/tafei_mall/model/vo/UserRegistration.java
package cn.edu.xidian.tafei_mall.model.vo;

import lombok.Data;



@Data
public class UserRegistration {
    private String username;
    private String password;
    private String email;

    // Getters and Setters
}